% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_player_idlist}
\alias{all_player_idlist}
\title{all player id list}
\format{A data frame with 1194 rows and 5 variables:
\describe{
\item{name}{name, Player name}
\item{namecode}{namecode, check same name}
\item{id}{id, player id}
\item{season}{season, play season}
\item{nickname}{nickname, classification same name}
}}
\source{
\url{https://github.com/LOPES-HUFS/KBO_Data_Wrangling}
}
\usage{
all_player_idlist
}
\description{
This data includes all KBO player IDs from 2010 to 2018.
}
\keyword{datasets}
