% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_statistic.R
\name{batter_statics}
\alias{batter_statics}
\title{Make Batter statistic}
\usage{
batter_statics(data)
}
\arguments{
\item{data}{KBO batter record data}
}
\value{
Batter's statistic
}
\description{
Function: Make Batter's statistic From the KBO data
}
\examples{
## internal function
#batter_statics(find_player(hanhwa_batter_2018,"이용규"))
}
