% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_cg_list.R
\name{bind_cg_list}
\alias{bind_cg_list}
\title{Make Complete game list}
\usage{
bind_cg_list(data, team)
}
\arguments{
\item{data}{KBO pitcher data}

\item{team}{KBO team name}
}
\value{
Complete game player and CG count data frame for all teams
}
\description{
Creating a complete game list from data.
}
\examples{
## internal function:
## This function creates Complete game data for the pitcher's boxscore.
# bind_cg_list(hanhwa_pitcher_2018,"한화")
}
