% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_cgs_list.R
\name{bind_cgs_list}
\alias{bind_cgs_list}
\title{Complete game Shutout rbind dataframe}
\usage{
bind_cgs_list(data, team)
}
\arguments{
\item{data}{KBO pitcher data}

\item{team}{KBO team name}
}
\value{
Complete game player and CGS count data frame for all teams
}
\description{
The calculated Complete game Shutout data frames are grouped into one data frame.
}
\examples{
## internal function:
## This function creates Complete game data for the pitcher's boxscore.
# bind_cgs_list(hanhwa_pitcher_2018,"한화")
}
