% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_player_id.R
\name{bind_player_id}
\alias{bind_player_id}
\title{Bind player id}
\usage{
bind_player_id(data, idlist)
}
\arguments{
\item{data}{KBO data}

\item{idlist}{The ID list contains the batter and pitcher ID list files in the player_id_list folder.}
}
\value{
Data with player id
}
\description{
This function creates a player id of the data.
}
\examples{
#notrun
#batter_id <- read.csv("./player_id_list/batter_id_list.csv",stringsAsFactors = F)
#id_2018 <- batter_id[substr(batter_id$date,1,4)==2018,]
#id_hanhwa_2018 <- id_2018[id_2018$away=="한화"|id_2018$home=="한화",]
#bind_player_id(hanhwa_batter_2018,"id_hanhwa_2018")
}
