% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_score.R
\name{box_score}
\alias{box_score}
\title{Make box score dataframe}
\usage{
box_score(jsondata, gameid)
}
\arguments{
\item{jsondata}{KBO data json file}

\item{gameid}{KBO game date and versus infomation}
}
\value{
gameid Shows the box score of the match that matches the information.
}
\description{
It shows the box score of the KBO game.
}
\examples{
## not run
## You can read the json file in the json_sample folder and use it.
## You can use the fromJSON function of the jsonlite package to read the json file.
## You can also get the KBO full game json file.
## You can import it from the "https://github.com/LOPES-HUFS/KBO_Data_Wrangling" project.
## jsondata <- jsonlite::fromJSON(./json_sample/Hanhwa_normalseason_2018.json)
# box_score(jsondata,names(jsondata)[1])
}
