% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtest.R
\name{boxscore_test}
\alias{boxscore_test}
\title{Inspection to distitinguish player names from team names}
\usage{
boxscore_test(name, boxscore, cg_list, cgs_list)
}
\arguments{
\item{name}{KBO pitcher name or KBO Team}

\item{boxscore}{The result value of the pitcher_statistics function}

\item{cg_list}{The result value of the bind_cg_list function}

\item{cgs_list}{The result value of the bind_cgs_list function}
}
\value{
Annual box score records for players or teams
}
\description{
Creates a box score to distinguish whether the name is a team name or not.
}
\examples{
## internal function
#name_test("한화",boxscore,cg_list,cgs_list)
}
