% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_game.R
\name{cg_calculate}
\alias{cg_calculate}
\title{Complete game count}
\usage{
cg_calculate(data, team)
}
\arguments{
\item{data}{KBO pitcher data}

\item{team}{KBO team name}
}
\value{
List of players who achieved Complete game
}
\description{
Creating a list of players who have achieve Complete game on the team.
}
\examples{
cg_calculate(hanhwa_pitcher_2018,"한화")
}
