% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_game_shutout.R
\name{cgs_calculate}
\alias{cgs_calculate}
\title{Complete Game Shutout count}
\usage{
cgs_calculate(data, team)
}
\arguments{
\item{data}{KBO pitcher data}

\item{team}{KBO team name}
}
\value{
List of players who achieved Complete Game Shutout
}
\description{
Creating a list of players who have achieve Complete Game Shutout on the team.
}
\examples{
cgs_calculate(hanhwa_pitcher_2018,"한화")
}
