% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_test.R
\name{date_test}
\alias{date_test}
\title{yearly and monthly argument test}
\usage{
date_test(data, name, id, yearly, monthly, recordname)
}
\arguments{
\item{data}{Default is hanhwa's batter data, You can include the entire KBO data if you wish.}

\item{name}{The name of the batter or the team you want}

\item{id}{Default is NULL, but you can put in Player id of player with same name}

\item{yearly}{Default is NULL, but you can put in the year}

\item{monthly}{Default is NULL, but you can put in the month}

\item{recordname}{The name of the record you want}
}
\value{
Record dataframe by period according to test
}
\description{
Tests whether the two arguments are null.
}
\examples{
date_test(hanhwa_batter_2018,"이용규",NULL,NULL,NULL,avg_formula)
}
