% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earned_run_formula.R
\name{er_formula}
\alias{er_formula}
\title{Earned run calculate function}
\usage{
er_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player ER
}
\description{
Calculate ER of player
}
\examples{
## internal function
#er_formula(hanhwa_pitcher_2018)
}
