% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_count.R
\name{factor_count}
\alias{factor_count}
\title{Batting record count}
\usage{
factor_count(data, record_factor)
}
\arguments{
\item{data}{KBO batter inning data}

\item{record_factor}{This is the category of batting records to find, and it should be a string type.}
}
\value{
Sum of the batter records in every innings
}
\description{
This function counts the number of batting records you put in.
}
\examples{
lee <- find_player(hanhwa_batter_2018,"이용규",NULL)
inning <- c('one','two','three','four','five','six','seven','eight','nine','ten','eleven','twelve')
inning_data <- subset(lee,select=inning)
factor_count(data=inning_data,record_factor="10")
}
