% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_etcname.R
\name{find_etcname}
\alias{find_etcname}
\title{Find etc name}
\usage{
find_etcname(jsondata, gameid, name1, name2, name3)
}
\arguments{
\item{jsondata}{KBO data json file}

\item{gameid}{KBO game date and versus infomation}

\item{name1}{find item1}

\item{name2}{find item2}

\item{name3}{find item3}
}
\value{
find items etc info
}
\description{
find etc name.
}
\examples{
## internal function
## not run
# jsondata <- jsonlite::fromJSON("./json_sample/Hanhwa_normalseason_2018.json")
# find_etcname(jsondata,names(jsondata)[1],"구장","관중","경기시간")
}
