% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_team.R
\name{find_only_team}
\alias{find_only_team}
\title{Extracting only one team data}
\usage{
find_only_team(data, teamname)
}
\arguments{
\item{data}{KBO data with players from different teams}

\item{teamname}{The teamname you want to find}
}
\value{
The batter data only the team you want to find
}
\description{
find_only_team function creates data that contains only the teams you want.
}
