% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_player_id.R
\name{find_player_id}
\alias{find_player_id}
\title{Extracting player id}
\usage{
find_player_id(playername)
}
\arguments{
\item{playername}{The playername you want to find}
}
\value{
the player id that you want to find
}
\description{
This function finds the id of the player you want.
}
\examples{
find_player_id("이용규")
}
