% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_team.R
\name{find_team}
\alias{find_team}
\title{Extracting team data}
\usage{
find_team(data, team)
}
\arguments{
\item{data}{KBO data with playername}

\item{team}{The team you want to find}
}
\value{
The batter data only the team you want to find
}
\description{
find_team function creates data that contains team you want.
}
\examples{
find_team(hanhwa_batter_2018,"한화")
}
