% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_into_double_play_formula.R
\name{gdp_formula}
\alias{gdp_formula}
\title{Ground into double play calculate function}
\usage{
gdp_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
the number of Ground into double play
}
\description{
Count Ground into double play of player
}
\examples{
## internal function
#gdp_formula(hanhwa_batter_2018)
}
