% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_id.R
\name{get_player_id}
\alias{get_player_id}
\title{Get player id}
\usage{
get_player_id(data, idlist)
}
\arguments{
\item{data}{KBO data}

\item{idlist}{The ID list contains the batter and pitcher ID list files in the player_id_list folder.}
}
\value{
Data with player id
}
\description{
This function creates a player id for the players in the data.
}
\examples{
#notrun
#batter_id  <- read.csv("./player_id_list/batter_id_full.csv",stringsAsFactors = F)
#id_2018<- batter_id[substr(batter_id$date,1,4)==2018,]
#id_hanhwa_2018<- id_2018[id_2018$away=="한화"|id_2018$home=="한화",]
#get_player_id(hanhwa_batter_2018,"id_hanhwa_2018")
}
