% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_allowed_formula.R
\name{ha_formula}
\alias{ha_formula}
\title{Hit allowed calculate function}
\usage{
ha_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player HA
}
\description{
Calculate HA of player
}
\examples{
## internal function
#ha_formula(hanhwa_pitcher_2018)
}
