% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hanhwa_batter_2018}
\alias{hanhwa_batter_2018}
\title{2018 season hanhwa batters data}
\format{A data frame with 3580 rows and 23 variables:
\describe{
\item{one}{one, 1 inning}
\item{two}{two, 2 inning}
\item{three}{three, 3 inning}
\item{four}{four, 4 inning}
\item{five}{five, 5 inning}
\item{six}{six, 6 inning}
\item{seven}{seven, 7 inning}
\item{eight}{eight, 8 inning}
\item{nine}{nine, 9 inning}
\item{ten}{ten, 10 inning}
\item{eleven}{eleven, 11 inning}
\item{twelve}{twelve, 12 inning}
\item{name}{name, PlayerNAME}
\item{r}{r, Runs Scored}
\item{h}{h, Hit}
\item{ab}{ab, AB: At Bat}
\item{rbi}{rbi, Run Batted In}
\item{team}{team, Player team}
\item{position}{position, Batter Position}
\item{date}{date, Game Date}
\item{away}{away, Awayteam}
\item{home}{home, Awayteam}
\item{doubleheader}{doubleheader, Whether it is a double-header game or not}
 }}
\source{
\url{https://github.com/LOPES-HUFS/KBO_Data_Wrangling}
}
\usage{
hanhwa_batter_2018
}
\description{
This data includes Hanhwa and the opponent's batter record in 2018.
}
\keyword{datasets}
