% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hanhwa_pitcher_2018}
\alias{hanhwa_pitcher_2018}
\title{2018 season hanhwa pitchers data}
\format{A data frame with 3580 rows and 21 variables:
\describe{
\item{date}{date, Game Date}
\item{away}{away, Awayteam}
\item{home}{home, Awayteam}
\item{doubleheader}{doubleheader, Whether it is a double-header game or not}
\item{name}{name, PlayerNAME}
\item{team}{team, Player team}
\item{position}{position, Pitcher Position}
\item{mound}{mound, Information on innings and batter numbers when a pitcher plays}
\item{win}{win, Whether or not it was a winning pitcher}
\item{lose}{lose, Whether or not it was a losing pitcher}
\item{draw}{draw, If the pitcher has a draw}
\item{hld}{hld, Hold record}
\item{sv}{sv, Save record}
\item{inning}{inning, Throw more than one inning}
\item{restinning}{restinning, Throw less than one inning}
\item{r}{r, The pitcher's run}
\item{er}{er, The pitcher's earned run.}
\item{k}{k, Pitcher's strikeout number}
\item{bb}{bb, Number of base on balls and number of hit by pitch}
\item{ab}{ab, AB: At Bat}
\item{tbf}{tbf, TBF:Total Batters Faced}
}}
\source{
\url{https://github.com/LOPES-HUFS/KBO_Data_Wrangling}
}
\usage{
hanhwa_pitcher_2018
}
\description{
This data includes Hanhwa and the opponent's pitcher record in 2018.
}
\keyword{datasets}
