% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_bb_formaula.R
\name{hb_bb_formula}
\alias{hb_bb_formula}
\title{HB(Hit batsman)+BB+IBB calculate function}
\usage{
hb_bb_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player HB+BB+IBB
}
\description{
Calculate HB+BB+IBB of player
}
\examples{
## internal function
#pit_formula(hanhwa_pitcher_2018)
}
