% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_by_pitch_formula.R
\name{hbp_formula}
\alias{hbp_formula}
\title{Hit by pitch calculate function}
\usage{
hbp_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
the number of  Hit by pitch
}
\description{
count Hit by pitch of player
}
\examples{
## internal function
#hbp_formula(hanhwa_batter_2018)
}
