% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hold_formula.R
\name{hld_formula}
\alias{hld_formula}
\title{HOLD calculate function}
\usage{
hld_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player HLD
}
\description{
Calculate HLD of player
}
\examples{
## internal function
#hld_formula(hanhwa_pitcher_2018)
}
