% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_formula.R
\name{homerun_formula}
\alias{homerun_formula}
\title{Homerun calculate function}
\usage{
homerun_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
the number of Homerun
}
\description{
count Homerun of player
}
\examples{
## internal function
#homerun_formula(hanhwa_batter_2018)
}
