% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_9_formula.R
\name{hr_9_formula}
\alias{hr_9_formula}
\title{Homerun / 9 calculate function}
\usage{
hr_9_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player HR / 9
}
\description{
Calculate HR / 9 of player
}
\examples{
## internal function
#hr_9_formula(hanhwa_pitcher_2018[hanhwa_pitcher_2018$team=="한화",])
}
