% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_allowed_formula.R
\name{hra_formula}
\alias{hra_formula}
\title{Homerun allowed calculate function}
\usage{
hra_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player HRA
}
\description{
Calculate HRA of player
}
\examples{
## internal function
#hra_formula(hanhwa_pitcher_2018)
}
