% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intentional_base_on_balls_formula.R
\name{ibb_formula}
\alias{ibb_formula}
\title{Intentional Base on Balls count}
\usage{
ibb_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
IBB from data
}
\description{
Counting Intentional Base on Balls from data.
}
\examples{
ibb_formula(hanhwa_batter_2018)
}
