% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innings_pitched_formula.R
\name{ip_formula}
\alias{ip_formula}
\title{Innings Pitched formula}
\usage{
ip_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
IP from data
}
\description{
Calculating Innings Pitched from data.
}
\examples{
ip_formula(hanhwa_pitcher_2018)
}
