% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondata2df.R
\name{json2batterdf}
\alias{json2batterdf}
\title{json file to dataframe}
\usage{
json2batterdf(jsondata, gameid)
}
\arguments{
\item{jsondata}{KBO data json file}

\item{gameid}{KBO game date and versus infomation}
}
\value{
One game is made up of data frames.
}
\description{
A function that creates a json file as a data frame.
}
\examples{
## You can read the json file in the json_sample folder and use it as an argument.
## You can use the fromJSON function of the jsonlite package to read the json file.
## You can also get the KBO full game json file by getting the "https://github.com/LOPES-HUFS/KBO_Data_Wrangling" project and working the code in the notebook file.
## If you get a json file through one of the above methods, you can open it and use it as an argument to the function.
## The gameid argument can be obtained by opening your json file and using the default function names.
## jsondata <- jsonlite::fromJSON(./json_sample/Hanhwa_normalseason_2018.json)
# json2batterdf(jsondata,names(jsondata)[1])
}
