% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_9_formula.R
\name{k_9_formula}
\alias{k_9_formula}
\title{K/9 formula}
\usage{
k_9_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
K/9 from data
}
\description{
Calculating K/9 from data.
}
\examples{
k_9_formula(hanhwa_pitcher_2018)
}
