% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strikeout_formula.R
\name{k_formula}
\alias{k_formula}
\title{K calculate function}
\usage{
k_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player K
}
\description{
Calculate K of player
}
\examples{
## internal function
#k_formula(hanhwa_pitcher_2018)
}
