% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_batter_of_teams.R
\name{make_all_batter_data}
\alias{make_all_batter_data}
\title{All batters' data for plotting}
\usage{
make_all_batter_data(data, team1, team2, yearly, monthly)
}
\arguments{
\item{data}{KBO batter data}

\item{team1}{The first of the team you want to compare}

\item{team2}{The second of the team you want to compare}

\item{yearly}{The default value is 2018. The role of determining the annual period of data}

\item{monthly}{The role of determining the monthly period of data}
}
\value{
Combined data all the batter data from two teams
}
\description{
Function: Make data of all batters from two teams for plotting
}
\examples{
make_all_batter_data(hanhwa_batter_2018,"한화","LG",2018,NULL)
}
