% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_all_pitcher_data.R
\name{make_all_pitcher_data}
\alias{make_all_pitcher_data}
\title{Make all pitchers of two team data for plotting}
\usage{
make_all_pitcher_data(data, team1, team2, yearly = 2018,
  monthly = NULL)
}
\arguments{
\item{data}{KBO pitcher data}

\item{team1}{The first of the team you want to compare}

\item{team2}{The second of the team you want to compare}

\item{yearly}{The role of determining the annual period of data}

\item{monthly}{The role of determining the monthly period of data}
}
\value{
Combined two team datas of all pitchers
}
\description{
Function: Make all pitchers of two team data for plotting
}
\examples{
## internal function
#make_all_pitcher_data(KBO_full_data,"한화","LG",2018,NULL)
}
