% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_data.R
\name{make_batter_data}
\alias{make_batter_data}
\title{Combines data from two batter boxscore}
\usage{
make_batter_data(data, playername1, playername2, yearly = NULL,
  monthly = NULL)
}
\arguments{
\item{data}{KBO batter data}

\item{playername1}{The first of the players you want to compare}

\item{playername2}{The second of the players you want to compare}

\item{yearly}{Determines the yearly period of the data.}

\item{monthly}{Determines the monthly period of the data.}
}
\value{
Combined data of two player boxscore
}
\description{
Function: Create data that combines data from two players
}
\examples{
## internal function
#make_batter_data(hanhwa_batter_2018,"이용규","정근우")
}
