% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_batter_data_4plot.R
\name{make_batter_data_4plot}
\alias{make_batter_data_4plot}
\title{Combine two batter boxscores}
\usage{
make_batter_data_4plot(data, playername1, playername2, yearly = NULL,
  monthly = NULL)
}
\arguments{
\item{data}{KBO batter data}

\item{playername1}{The first of the players you want to compare}

\item{playername2}{The second of the players you want to compare}

\item{yearly}{The role of determining the annual period of data}

\item{monthly}{The role of determining the monthly period of data}
}
\value{
Combined boxscore data for two players
}
\description{
Function: Creates data that combines two batters.
}
\examples{
make_batter_data_4plot(hanhwa_batter_2018,"이용규","정근우")
}
