% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pitcher_data_4plot.R
\name{make_pitcher_data_4plot}
\alias{make_pitcher_data_4plot}
\title{Combine two pitcher boxscores}
\usage{
make_pitcher_data_4plot(data, playername1, playername2, yearly = NULL,
  monthly = NULL)
}
\arguments{
\item{data}{KBO pitcher data}

\item{playername1}{The first of the players you want to compare}

\item{playername2}{The second of the players you want to compare}

\item{yearly}{The role of determining the annual period of data}

\item{monthly}{The role of determining the monthly period of data}
}
\value{
Box score data combined with two players' box scores
}
\description{
Function: Creates data that combines two pitchers.
}
\examples{
make_pitcher_data_4plot(hanhwa_pitcher_2018,"정우람","이태양")
}
