% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot4pitcher_data.R
\name{make_plot4pitcher_data}
\alias{make_plot4pitcher_data}
\title{Combine two pitcher boxscores}
\usage{
make_plot4pitcher_data(data, playername1, playername2, yearly = NULL,
  monthly = NULL)
}
\arguments{
\item{data}{KBO pitcher data}

\item{playername1}{The first of the players you want to compare}

\item{playername2}{The second of the players you want to compare}

\item{yearly}{The role of determining the annual period of data}

\item{monthly}{The role of determining the monthly period of data}
}
\value{
Combined boxscore data for two players
}
\description{
Function: Creates data that combines two pitchers.
}
\examples{
make_plot4pitcher_data(hanhwa_pitcher_2018,"정우람","이태양")
}
