% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_data.R
\name{make_plot_data}
\alias{make_plot_data}
\title{Make plotting data}
\usage{
make_plot_data(data, playername1, playername2, yearly = NULL,
  monthly = NULL)
}
\arguments{
\item{data}{KBO batter data}

\item{playername1}{The first of the players you want to compare}

\item{playername2}{The second of the players you want to compare}

\item{yearly}{The role of determining the annual period of data}

\item{monthly}{The role of determining the monthly period of data}
}
\value{
Combined data of two players
}
\description{
Function: Make Batter's Data for plotting
}
\examples{
## internal function
#make_plot_data(hanhwa_batter_2018,"이용규","정근우")
}
