% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_test.R
\name{name_test}
\alias{name_test}
\title{Check name tag}
\usage{
name_test(name)
}
\arguments{
\item{name}{KBO player name or KBO Team name}
}
\value{
name tag
}
\description{
This function identifies whether 'name' parameter is a team name or a player name.
}
\examples{
## internal function
#name_test("한화")
}
