% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_base_percentage_formula.R
\name{obp_formula}
\alias{obp_formula}
\title{On-base percentage calculate function}
\usage{
obp_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's OBP
}
\description{
OBP calculation in the data
}
\examples{
obp_formula(hanhwa_batter_2018)
}
