% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onebase_formula.R
\name{onebase_formula}
\alias{onebase_formula}
\title{1B calculating function}
\usage{
onebase_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's 1B
}
\description{
Calculates the data's single hit.
}
\examples{
onebase_formula(hanhwa_batter_2018)
}
