% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneseason_compare_pitcher.R
\name{oneseason_compare_pitcher}
\alias{oneseason_compare_pitcher}
\title{Plot comparing two pitcher stats with one season}
\usage{
oneseason_compare_pitcher(data = kbodatamining::hanhwa_pitcher_2018,
  playername1, playername2, yearly = 2018, monthly = NULL)
}
\arguments{
\item{data}{Default data is hanhwa_pitcher_2018 You can put other kbo batter data if you want.}

\item{playername1}{Name of the pitcher you want to compare}

\item{playername2}{Name of the pitcher you want to compare}

\item{yearly}{The default value is NULL Possible values are a specific year.}

\item{monthly}{The default value is NULL Possible values are a specific month.}
}
\value{
Plots comparing two pitcher stats per year
}
\description{
This function draws a plot that compares two pitchers based on one season batter data.
This function has an option for yearly and monthly.
default: Based on the 2018 season, the month argument is NULL
If you put specific year in the yearly parameter, draws a batting comparison plot for the matched year.
If you put specific month in the monthly parameter,
draws a pitching comparison plot for the matched 2018 year and specific month.
If you put specific year and month in two parameters,
draws a pitching comparison plot for the matched specific year and month.
If you are using mac and Hangul fonts are broken, type `theme_gray (base_family = "AppleGothic"))`
}
\examples{
## default: yearly=2018, monthly=NULL
oneseason_compare_pitcher(hanhwa_pitcher_2018,"정우람","이태양",yearly=2018)
## yearly=2018, monthly="05"
oneseason_compare_pitcher(hanhwa_pitcher_2018,"정우람","이태양",yearly=2018,monthly="05")
}
