% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneyear_batter_plot.R
\name{oneyear_batter_plot}
\alias{oneyear_batter_plot}
\title{Year batter data plotting}
\usage{
oneyear_batter_plot(data)
}
\arguments{
\item{data}{two batter player's merged data}
}
\value{
A plot comparing records
}
\description{
plotting one season batter data to compare
}
\examples{
## internal function
#oneyear_batter_plot(make_plot_data(hanhwa_batter_2018,"이용규","정근우"))
}
