% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneyear_pitcher_plot.R
\name{oneyear_pitcher_plot}
\alias{oneyear_pitcher_plot}
\title{Year pitcher data plotting}
\usage{
oneyear_pitcher_plot(data)
}
\arguments{
\item{data}{two pitcher player's merged data}
}
\value{
A plot comparing records
}
\description{
plotting one season pitcher data to compare
}
\examples{
## internal function
#oneyear_pitcher_plot(make_pitcher_data(hanhwa_pitcher_2018,"정우람","이태양"))
}
