% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneyear_plotting.R
\name{oneyear_plotting}
\alias{oneyear_plotting}
\title{Year data plotting}
\usage{
oneyear_plotting(data)
}
\arguments{
\item{data}{two batter player's merged data}
}
\value{
A plot comparing four records
}
\description{
plotting year batter data
}
\examples{
## internal function
#oneyear_plotting(make_plot_data(hanhwa_batter_2018,"이용규","정근우"))
}
