% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_appearance_formula.R
\name{pa_formula}
\alias{pa_formula}
\title{The number of Plate appearance calculate function}
\usage{
pa_formula(data)
}
\arguments{
\item{data}{KBO batter inning data}
}
\value{
player PA
}
\description{
Calculate PA of player
}
\examples{
## internal function
#pit_formula(hanhwa_batter_2018)
}
