% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitches_thrown_formula.R
\name{pit_formula}
\alias{pit_formula}
\title{The number of Pitches thrown calculate function}
\usage{
pit_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player PIT
}
\description{
Calculate PIT of player
}
\examples{
## internal function
#pit_formula(hanhwa_pitcher_2018)
}
