% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_boxscore.R
\name{pitcher_boxscore}
\alias{pitcher_boxscore}
\title{The box score of the pitcher player's period}
\usage{
pitcher_boxscore(data = kbodatamining::hanhwa_pitcher_2018, name,
  yearly = NULL, monthly = NULL)
}
\arguments{
\item{data}{Default data is hanhwa_pitcher_2018 You can put other kbo pitcher data if you want.}

\item{name}{The name of the pitcher you want}

\item{yearly}{The default value is NULL Possible values are a specific year.}

\item{monthly}{The default value is NULL Possible values are a specific month.}
}
\value{
The box score for the year, month, or period put in is created.
}
\description{
This function creates a box score for the pitcher for the year or period.
This function has an option for yearly and monthly.
default: The player's annual box score is created.
If you put a year in the yearly parameter, it calculates the corresponding year box score.
If you put a month in the monthly parameter, it calculates the corresponding yearly-month box score.
If you put a year and a month in two parameters, the box score for the specific time you put in is created.
}
\examples{
## default: yearly=NULL,monthly=NULL
pitcher_boxscore(hanhwa_pitcher_2018,"정우람")
## yearly=2018, monthly=NULL
pitcher_boxscore(hanhwa_pitcher_2018,"정우람",yearly=2018)
## yearly=NULL, monthly="05"
pitcher_boxscore(hanhwa_pitcher_2018,"정우람",monthly="05")
## yearly=2018, monthly="05"
pitcher_boxscore(hanhwa_pitcher_2018,"정우람",yearly=2018,monthly="05")
}
