% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitcher_statistic.R
\name{pitcher_statistics}
\alias{pitcher_statistics}
\title{Make Pitcher statistic}
\usage{
pitcher_statistics(data)
}
\arguments{
\item{data}{KBO Pitcher data}
}
\value{
Statistics of Pitcher
}
\description{
Function: Make Pitcher's statistic From the KBO Pitcher data
}
\examples{
## internal function:This function creates statistical data for the pitcher's boxscore.
#pitcher_statistics(find_player(hanhwa_pitcher_2018,"정우람",NULL))
}
