% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_3B.R
\name{player_3B}
\alias{player_3B}
\title{Calculating batter's Triples}
\usage{
player_3B(data = kbodatamining::hanhwa_batter_2018, playername,
  yearly = NULL, monthly = NULL)
}
\arguments{
\item{data}{Default data is hanhwa_batter_2018 You can put other kbo batter data if you want.}

\item{playername}{The name of the batter you want.}

\item{yearly}{The default value is NULL Possible values are a specific year.}

\item{monthly}{The default value is NULL Possible values are a specific month.}
}
\value{
The Triples for the year, month, or period put in is calculated.
}
\description{
This function calculates The Triples for the specified player.
This function has an option for yearly and monthly.
default: the annualy Triples is calculated
If you put a year in the yearly parameter, it calculates the corresponding year Triples.
If you put a month in the monthly parameter, it calculates the corresponding yearly-month Triples.
If you put a year and a month in two parameters, the Triples for the specific time you put in is calculated.
}
\examples{
## default
player_3B(hanhwa_batter_2018,"이용규")
## yearly=2018, monthly=NULL
player_3B(hanhwa_batter_2018,"이용규",yearly=2018)
## yearly=NULL, monthly="05"
player_3B(hanhwa_batter_2018,"이용규",monthly="05")
## yearly=2018, monthly="05"
player_3B(hanhwa_batter_2018,"이용규",yearly=2018,monthly="05")
}
