% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_avg.R
\name{player_avg}
\alias{player_avg}
\title{Calculate batting average}
\usage{
player_avg(data, playername, yearly = TRUE, monthly = FALSE)
}
\arguments{
\item{data}{KBO Batter data}

\item{playername}{The name of the batter you want}

\item{yearly}{The default value is TRUE Possible values are a specific year or a FALSE value.}

\item{monthly}{The default value is FALSE. Possible values are a specific month or a TRUE value.}
}
\value{
The batting average for the year, month, or period entered is calculated.
}
\description{
This function calculates the batting average for the specified player.
This function has an option for yearly and monthly.
default: the annual AVG is calculated
If only the monthly parameter is TRUE, the monthly AVG is calculated
If you put a year in the yearly parameter, it calculates the AVG for that year.
If you put a month in the monthly parameter, it calculates the corresponding monthly AVG.
If you put a year and a month in two parameter, the AVG for the time period you put in is calculated.
}
\examples{
## default
player_avg(hanhwa_batter_2018,"이용규")
## yearly=FALSE, monthly=TRUE
player_avg(hanhwa_batter_2018,"이용규",yearly=FALSE,monthly=TRUE)
## yearly=TRUE, monthly=TRUE #error case
player_avg(hanhwa_batter_2018,"이용규",yearly=TRUE,monthly=TRUE)
## yearly=2018, monthly=FALSE
player_avg(hanhwa_batter_2018,"이용규",yearly=2018)
#' ## yearly=2018, monthly=TRUE
player_avg(hanhwa_batter_2018,"이용규",yearly=2018,monthly=TRUE)
## yearly=TRUE, monthly=05
player_avg(hanhwa_batter_2018,"이용규",yearly=TRUE,monthly=05)
## yearly=2018, monthly=05
player_avg(hanhwa_batter_2018,"이용규",yearly=2018,monthly=05)
}
