% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_disposal.R
\name{player_disposal}
\alias{player_disposal}
\title{Process the player's complete game record}
\usage{
player_disposal(data, boxscore, name)
}
\arguments{
\item{data}{cg_list or cgs_list (The result value of bind_cg_list or bind_cgs_list)}

\item{boxscore}{The result value of the pitcher_statistics function}

\item{name}{player name}
}
\value{
boxscore with CG or CGS
}
\description{
Process for creating player box scores
}
\examples{
## internal function
#player_disposal(cg_list,boxscore)
}
