% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_ER.R
\name{player_ER}
\alias{player_ER}
\title{Count yearly-monthly Earned run}
\usage{
player_ER(data = kbodatamining::hanhwa_pitcher_2018, playername,
  yearly = NULL, monthly = NULL)
}
\arguments{
\item{data}{Default data is hanhwa_pitcher_2018, You can put other KBO pitcher data if you want.}

\item{playername}{The name of the KBO pitcher you want.}

\item{yearly}{The default value is NULL Possible values are a specific year.}

\item{monthly}{The default value is NULL Possible values are a specific month.}
}
\value{
The ER for the year, month, or period put in is calculated.
}
\description{
This function calculates the ER for the specified player.
This function has an option for yearly and monthly.
default: the annualy ER is calculated
If you put a year in the yearly parameter, it calculates the corresponding year ER.
If you put a month in the monthly parameter, it calculates the corresponding yearly-month ER.
If you put a year and a month in two parameters, the ER for the specific time you put in is calculated.
}
\examples{
## default
player_ER(hanhwa_pitcher_2018,"송은범")
## yearly=2018, monthly=NULL
player_ER(hanhwa_pitcher_2018,"송은범",yearly=2018)
## yearly=NULL, monthly="05"
player_ER(hanhwa_pitcher_2018,"송은범",monthly="05")
## yearly=2018, monthly="05"
player_ER(hanhwa_pitcher_2018,"송은범",yearly=2018,monthly="05")
}
