% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_OBP.R
\name{player_OBP}
\alias{player_OBP}
\title{Calculating batter's On-base percentage}
\usage{
player_OBP(data = kbodatamining::hanhwa_batter_2018, playername,
  yearly = NULL, monthly = NULL)
}
\arguments{
\item{data}{Default data is hanhwa_batter_2018 You can put other kbo batter data if you want.}

\item{playername}{The name of the batter you want.}

\item{yearly}{The default value is NULL Possible values are a specific year.}

\item{monthly}{The default value is NULL Possible values are a specific month.}
}
\value{
The On-base percentage for the year, month, or period put in is calculated.
}
\description{
This function calculates the On-base percentage for the specified player.
This function has an option for yearly and monthly.
default: the annualy On-base percentage is calculated
If you put a year in the yearly parameter, it calculates the corresponding year On-base percentage.
If you put a month in the monthly parameter, it calculates the corresponding yearly-monthOn-base percentage.
If you put a year and a month in two parameters, the On-base percentage for the specific time you put in is calculated.
}
\examples{
## default
player_OBP(hanhwa_batter_2018,"이용규")
## yearly=2018, monthly=NULL
player_OBP(hanhwa_batter_2018,"이용규",yearly=2018)
## yearly=NULL, monthly="05"
player_OBP(hanhwa_batter_2018,"이용규",monthly="05")
## yearly=2018, monthly="05"
player_OBP(hanhwa_batter_2018,"이용규",yearly=2018,monthly="05")
}
