% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_WPCT.R
\name{player_WPCT}
\alias{player_WPCT}
\title{Count yearly-monthly pitcher's winning percentage}
\usage{
player_WPCT(data = kbodatamining::hanhwa_pitcher_2018, playername,
  yearly = NULL, monthly = NULL)
}
\arguments{
\item{data}{Default data is hanhwa_pitcher_2018, You can put other KBO pitcher data if you want.}

\item{playername}{The name of the KBO pitcher you want.}

\item{yearly}{The default value is NULL Possible values are a specific year.}

\item{monthly}{The default value is NULL Possible values are a specific month.}
}
\value{
The WPCT for the year, month, or period put in is calculated.
}
\description{
This function calculates the WPCT for the specified player.
This function has an option for yearly and monthly.
default: the annualy WPCT is calculated
If you put a year in the yearly parameter, it calculates the corresponding year WPCT.
If you put a month in the monthly parameter, it calculates the corresponding yearly-month WPCT.
If you put a year and a month in two parameters, the HA for the specific time you put in is calculated.
}
\examples{
## default
player_WPCT(hanhwa_pitcher_2018,"송은범")
## yearly=2018, monthly=NULL
player_WPCT(hanhwa_pitcher_2018,"송은범",yearly=2018)
## yearly=NULL, monthly="05"
player_WPCT(hanhwa_pitcher_2018,"송은범",monthly="05")
## yearly=2018, monthly="05"
player_WPCT(hanhwa_pitcher_2018,"송은범",yearly=2018,monthly="05")
}
