% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreboard.R
\name{scoreboard}
\alias{scoreboard}
\title{Make scoreboard dataframe}
\usage{
scoreboard(jsondata, gameid)
}
\arguments{
\item{jsondata}{KBO data json file}

\item{gameid}{KBO game date and versus infomation}
}
\value{
the baseball match of gameid appears in scoreboard dataframe.
}
\description{
It shows the scoreboard of the KBO game.
}
\examples{
## not run
# jsondata <- jsonlite::fromJSON("./json_sample/Hanhwa_normalseason_2018.json")
# scoreboard(jsondata,names(jsondata)[1])
}
