% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacrifice_fly_formula.R
\name{sf_formula}
\alias{sf_formula}
\title{Sacrifice hit calculate function}
\usage{
sf_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
the number of Sacrifice fly
}
\description{
count Sacrifice fly of player
}
\examples{
## internal function
#sf_formula(hanhwa_batter_2018)
}
