% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacrifice_hit_formula.R
\name{sh_formula}
\alias{sh_formula}
\title{Sacrifice Hit formula}
\usage{
sh_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
Sacrifice Hit from data
}
\description{
Calculating Sacrifice Hit from data
}
\examples{
sh_formula(hanhwa_batter_2018)
}
