% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugging_average_formula.R
\name{slg_formula}
\alias{slg_formula}
\title{Slugging average calculate function}
\usage{
slg_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's Slugging average
}
\description{
Calculate Slugging average in the data
}
\examples{
slg_formula(hanhwa_batter_2018)
}
