% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_formula.R
\name{so_formula}
\alias{so_formula}
\title{StrikeOut(SO) calculate formula}
\usage{
so_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's SO
}
\description{
This function calculates the strikeout in which the baseball player was hit in the data.
}
\examples{
so_formula(hanhwa_batter_2018)
}
