% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_col.R
\name{sum_col}
\alias{sum_col}
\title{Col sum function}
\usage{
sum_col(data, col)
}
\arguments{
\item{data}{KBO data}

\item{col}{Selected col name}
}
\value{
Sum of selected columns
}
\description{
Select and sum one of the columns in the data
}
\examples{
sum_col(hanhwa_pitcher_2018,"er")
}
