% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_formula.R
\name{sv_formula}
\alias{sv_formula}
\title{Save calculate function}
\usage{
sv_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player SV
}
\description{
Calculate SV of player
}
\examples{
## internal function
#sv_formula(hanhwa_pitcher_2018)
}
