% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_batters_faced_formula.R
\name{tbf_formula}
\alias{tbf_formula}
\title{Total Batters Faced calculate function}
\usage{
tbf_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player TBF
}
\description{
Calculate TBF of player
}
\examples{
## internal function
#tbf_formula(hanhwa_pitcher_2018)
}
