% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_disposal.R
\name{team_disposal}
\alias{team_disposal}
\title{Process the team's complete game record}
\usage{
team_disposal(data, boxscore, teamname)
}
\arguments{
\item{data}{cg_list or cgs_list (The result value of cg_calculate or cgs_calculate)}

\item{boxscore}{The result value of the pitcher_statistics function}

\item{teamname}{teamname}
}
\value{
yearly CG or CGS
}
\description{
Process for creating team box scores
}
\examples{
## internal function - not run
# boxscore <- date_test(data=hanhwa_pitcher_2018,teamname="한화",id=NULL,2018,NULL,pitcher_statistics)
# team_disposal(cg_calculate(hanhwa_pitcher_2018,"한화"),boxscore,"한화")
}
