% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_pitcher_boxscore.R
\name{team_pitcher_boxscore}
\alias{team_pitcher_boxscore}
\title{The pitcher box score of the team}
\usage{
team_pitcher_boxscore(data = kbodatamining::hanhwa_pitcher_2018,
  teamname, yearly = NULL, monthly = NULL)
}
\arguments{
\item{data}{'Hanhwa_pitcher_2018' is default data. You can put another kbo pitcher data if you want.}

\item{teamname}{The name of team you want}

\item{yearly}{The default value is NULL. Possible value is a specific year.}

\item{monthly}{The default value is NULL. Possible value is a specific month. The monthly parameter is a string type.}
}
\value{
The team pitching box score of the year or month put in is created.
}
\description{
This function creates annual or periodical box score of the team pitching
This function has options called 'yearly' or 'monthly'.
default: The team's annual box score is created.
If you put a year in the yearly parameter, it calculates the box score of corresponding year.
If you put a month in the monthly parameter, it calculates the corresponding 'monthly' box score of every year.
If you put a year and a month in each parameters, the box score of corresponding month you put in is created.
}
\examples{
## Example default: yearly=NULL, monthly=NULL
team_pitcher_boxscore(data=hanhwa_pitcher_2018,teamname="한화")
## yearly=2018, monthly=NULL
team_pitcher_boxscore(data=hanhwa_pitcher_2018,teamname="한화",yearly=2018)
## yearly=NULL, monthly="05"
team_pitcher_boxscore(data=hanhwa_pitcher_2018,teamname="한화",monthly="05")
## yearly=2018, monthly="05"
team_pitcher_boxscore(data=hanhwa_pitcher_2018,teamname="한화",yearly=2018,monthly="05")
}
