% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_pitcher_compare_plot.R
\name{team_pitcher_compare_plot}
\alias{team_pitcher_compare_plot}
\title{Compares the hits of two teams}
\usage{
team_pitcher_compare_plot(data)
}
\arguments{
\item{data}{two teams of batter's merged data}
}
\value{
A plot comparing records(ERA,K(SO),OBA,P/IP)
}
\description{
A plotting function that compares the hits of two teams
}
\examples{
## internal function
#team_pitcher_compare_plot(make_all_pitcher_data(KBO_full_data,"한화","LG",2018,NULL))
}
