% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threebase_formula.R
\name{threebase_formula}
\alias{threebase_formula}
\title{3B calculate function}
\usage{
threebase_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
the number of 3B
}
\description{
count Triples of player
}
\examples{
## internal function
#threebase_formula(hanhwa_batter_2018)
}
