% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twobase_formula.R
\name{twobase_formula}
\alias{twobase_formula}
\title{2B count}
\usage{
twobase_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
2B from data
}
\description{
Counting Double from data
}
\examples{
twobase_formula(hanhwa_batter_2018)
}
