% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winning_percentage_formula.R
\name{wpct_formula}
\alias{wpct_formula}
\title{Winning Percentage(WPCT) formula}
\usage{
wpct_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
WPCT from data
}
\description{
Calculating WPCT from data
}
\examples{
wpct_formula(hanhwa_pitcher_2018)
}
