% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_month_record.R
\name{year_month_record}
\alias{year_month_record}
\title{year month record show}
\usage{
year_month_record(data, playername, recordname, year, month)
}
\arguments{
\item{data}{Default is hanhwa's batter data, You can include the entire KBO data if you wish.}

\item{playername}{The name of the batter you want}

\item{recordname}{The name of the record you want}

\item{year}{The year you want to find}

\item{month}{The month you want to find}
}
\value{
show year-month batter record
}
\description{
It shows the year-month batter record.
}
\examples{
## internal function
#year_monthly_record(hanhwa_batter_2018,"이용규",avg_formula)
}
