% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_batter_plot.R
\name{yearly_batter_plot}
\alias{yearly_batter_plot}
\title{Yearly data plotting}
\usage{
yearly_batter_plot(data)
}
\arguments{
\item{data}{two batter player's merged data}
}
\value{
A plot comparing four records(G,AVG,OPS,BABIP)
}
\description{
plotting yearly batter data
}
\examples{
## internal function
#yearly_plotting(make_plot_data(hanhwa_batter_2018,"이용규","정근우"))
}
