% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_compare_pitcher.R
\name{yearly_compare_pitcher}
\alias{yearly_compare_pitcher}
\title{Plot comparing two pitcher stats}
\usage{
yearly_compare_pitcher(data, playername1, playername2, monthly = NULL)
}
\arguments{
\item{data}{KBO pitcher data over 2 seasons}

\item{playername1}{Name of the pitcher you want to compare}

\item{playername2}{Name of the pitcher you want to compare}

\item{monthly}{The default value is NULL, Possible values are a specific month string such as "05"}
}
\value{
Four plots comparing the four records(G, AVG, OPS, BABIP) of the players entered
}
\description{
This function draws a plot that compares two pitcher based on the yearly pitcher data.
This function has an option monthly.
If you put in a specific month for monthly parameter, you will compare the pitchers with the put in month data
## You can refer to the Readme file to obtain the annual KBO data
}
\examples{
## not run
## You can refer to the Readme file to obtain the annual KBO data
## default:monthly=NULL
## yearly_compare_pitcher(yearly_KBO_data,"정우람","송은범")
## monthly="05
## yearly_compare_pitcher(yearly_KBO_data,"정우람","이태양",monthly="05")
}
