% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_month_record.R
\name{yearly_month_record}
\alias{yearly_month_record}
\title{Annual record for the entered month}
\usage{
yearly_month_record(data, name, recordname, month)
}
\arguments{
\item{data}{KBO datas with at least one year of data}

\item{name}{The name of the player or the team you want}

\item{recordname}{The name of the record you want}

\item{month}{The month you want to find}
}
\value{
the record of the month entered in every year of the entered player
}
\description{
It shows the record of the month entered in every year of the entered player
}
\examples{
## internal function
#yearly_month_record(hanhwa_batter_2018,"이용규",avg_formula,"05")
#yearly_month_record(hanhwa_batter_2018,"한화",avg_formula,"05")
}
