% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_pitcher_plot.R
\name{yearly_pitcher_plot}
\alias{yearly_pitcher_plot}
\title{Annual pitch record plot}
\usage{
yearly_pitcher_plot(data)
}
\arguments{
\item{data}{two pitcher player's merged data}
}
\value{
A plot comparing four records(G,ERA,K,IP)
}
\description{
Compare the four records of the two players in the data.
This function requires data that has a duration of at least two seasons.
}
\examples{
## internal function
#yearly_pitcher_plot(make_pitcher_data_4plot(Yearly_KBO_DATA,"정우람","송은범"))
}
