% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_formula.R
\name{avg_formula}
\alias{avg_formula}
\title{AVG formula}
\usage{
avg_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
AVG from data
}
\description{
Calculating AVG from data
}
\examples{
avg_formula(hanhwa_batter_2018)
}
