% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_formula.R
\name{avg_formula}
\alias{avg_formula}
\title{avg calculate formula}
\usage{
avg_formula(h, ab)
}
\arguments{
\item{h}{hit record data}

\item{ab}{at bat record data}
}
\value{
calculate avg
}
\description{
only calculate function
}
\examples{
## internal function
#avg_formula(h,ab)
}
