% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_formula.R
\name{avg_formula}
\alias{avg_formula}
\title{AVG calculate formula}
\usage{
avg_formula(data)
}
\arguments{
\item{data}{batter record data}
}
\value{
calculate avg
}
\description{
only AVG calculate function
}
\examples{
## internal function
#avg_formula(hanhwa_batter_2018)
}
