% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batter_statistic.R
\name{batter_statistics}
\alias{batter_statistics}
\title{Make Batter statistic}
\usage{
batter_statistics(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
Batter's statistic
}
\description{
Function: Make Batter's statistic From the KBO batter data
}
\examples{
## internal function:This function creates statistical data for the batter's boxscore.
#batter_statistics(find_player(hanhwa_batter_2018,"이용규",NULL))
}
