% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_on_balls_formula.R
\name{bb_formula}
\alias{bb_formula}
\title{Base on balls count}
\usage{
bb_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
base on balls from data
}
\description{
Counting base on balls from data
}
\examples{
bb_formula(hanhwa_batter_2018)
}
