% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_on_balls_formula.R
\name{bb_formula}
\alias{bb_formula}
\title{Base on balls calculate function}
\usage{
bb_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
the number of base on balls
}
\description{
count base on balls of player
}
\examples{
## internal function
#bb_formula(hanhwa_batter_2018)
}
