% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_cg_list.R
\name{bind_cg_list}
\alias{bind_cg_list}
\title{Complete game rbind dataframe}
\usage{
bind_cg_list(data, team)
}
\arguments{
\item{data}{KBO pitcher data}

\item{team}{KBO team name}
}
\value{
Complete game player and CG count data frame for all teams
}
\description{
The calculated Complete game data frames are grouped into one data frame.
}
\examples{
bind_cg_list(hanhwa_pitcher_2018,"한화")
}
