% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtest.R
\name{boxscore_test}
\alias{boxscore_test}
\title{Makes the complete pitcher's box score.}
\usage{
boxscore_test(name, boxscore, cg_list, cgs_list)
}
\arguments{
\item{name}{KBO pitcher name or KBO Team}

\item{boxscore}{The result value of the pitcher_statistics function}

\item{cg_list}{The result value of the bind_cg_list function}

\item{cgs_list}{The result value of the bind_cgs_list function}
}
\value{
complete pitcher box score
}
\description{
Creating a complete box score with CG_list, CGS_list, and boxed score
}
\examples{
## internal function
## not run
#boxscore_test("한화",boxscore,cg_list,cgs_list)
}
