% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_game.R
\name{cg_calculate}
\alias{cg_calculate}
\title{Complete game calculate function}
\usage{
cg_calculate(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
Complete game player and cg count dataframe
}
\description{
Calculate Complete game
}
\examples{
## internal function
#cg_calculate(find_team(hanhwa_pitcher_2018,"한화"))
}
