% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_game_shutdown.R
\name{cgs_calculate}
\alias{cgs_calculate}
\title{Complete game Shutout calculate function}
\usage{
cgs_calculate(data, team)
}
\arguments{
\item{data}{KBO pitcher data}

\item{team}{KBO team name}
}
\value{
Complete game Shutout dataframe
}
\description{
Calculate Complete game Shutout
}
\examples{
cgs_calculate(hanhwa_pitcher_2018,"한화")
}
