% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_game_shutdown.R
\name{cgs_calculate}
\alias{cgs_calculate}
\title{Complete game Shutout calculate function}
\usage{
cgs_calculate(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
Complete game Shutout dataframe
}
\description{
Calculate Complete game Shutout
}
\examples{
## internal function
#cgs_calculate(find_team(hanhwa_pitcher_2018,"한화"))
}
