% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_test.R
\name{date_test}
\alias{date_test}
\title{yearly and monthly parameter test}
\usage{
date_test(data, name, id, yearly, monthly, recordname)
}
\arguments{
\item{data}{KBO batter data or KBO pitcher data}

\item{name}{The name of the batter or the team you want}

\item{id}{This is not a required parameter. The ID parameter is a numeric type.}

\item{yearly}{The default value is NULL. Possible value is a specific year.}

\item{monthly}{The default value is NULL. Possible value is a specific month.}

\item{recordname}{The name of the record you want}
}
\value{
Baseball record data created according to 'yearly', 'monthly' parameter
}
\description{
Tests whether the two parameters are null.
}
\examples{
## internal function
date_test(hanhwa_batter_2018,"이용규",NULL,NULL,NULL,avg_formula)
}
