% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earned_run_average.R
\name{era_formula}
\alias{era_formula}
\title{Earned run average calculate function}
\usage{
era_formula(data)
}
\arguments{
\item{data}{pitcher inning data}
}
\value{
player ERA
}
\description{
Calculate ERA of player
}
\examples{
## internal function
#era_formula(hanhwa_pitcher_2018)
}
