% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earned_run_average_formula.R
\name{era_formula}
\alias{era_formula}
\title{Earned Run Average count}
\usage{
era_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
ERA from data
}
\description{
Counting ERA from the data
}
\examples{
era_formula(hanhwa_pitcher_2018)
}
