% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_id.R
\name{get_player_id}
\alias{get_player_id}
\title{Get player id}
\usage{
get_player_id(data, idlist)
}
\arguments{
\item{data}{KBO batter data or KBO pitcher data.}

\item{idlist}{appropriate id list for your data}
}
\value{
Player data with player id
}
\description{
Matching player data and player id appropriately
}
\examples{
## not run
## please check README
#batter_id  <- read.csv("./player_id_list/batter_id_list.csv",stringsAsFactors = F)
#batter_id_2018 <- batter_id[substr(batter_id$date,1,4)==2018,]
#batter_id_hanhwa_2018 <- batter_id_2018[batter_id_2018$away=="한화"|batter_id_2018$home=="한화",]
#hanhwa_batter_2018 <- get_player_id(hanhwa_batter_2018,batter_id_hanhwa_2018)
}
