% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_into_double_play_formula.R
\name{gidp_formula}
\alias{gidp_formula}
\title{Ground Into Double Play count}
\usage{
gidp_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
GIDP from the data
}
\description{
Counting GIDP from the data
}
\examples{
gidp_formula(hanhwa_batter_2018)
}
