% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_into_double_play_formula.R
\name{gidp_formula}
\alias{gidp_formula}
\title{Ground into double play calculate function}
\usage{
gidp_formula(data)
}
\arguments{
\item{data}{KBO batter inning data}
}
\value{
The player's GIDP
}
\description{
Calculates the ground into double play record in the data.
}
\examples{
gidp_formula(hanhwa_batter_2018)
}
