% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_by_pitch_formula.R
\name{hbp_formula}
\alias{hbp_formula}
\title{Hit By Pitch count}
\usage{
hbp_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
Hit By Pitch from data
}
\description{
Counting Hit By Pitch from data
}
\examples{
hbp_formula(hanhwa_batter_2018)
}
