% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_formula.R
\name{homerun_formula}
\alias{homerun_formula}
\title{Homerun calculate function}
\usage{
homerun_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's Homerun
}
\description{
Calculates Homerun in the data.
}
\examples{
homerun_formula(hanhwa_batter_2018)
}
