% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_formula.R
\name{homerun_formula}
\alias{homerun_formula}
\title{Homerun count}
\usage{
homerun_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
Homerun from data
}
\description{
Counting Homerun from data
}
\examples{
homerun_formula(hanhwa_batter_2018)
}
