% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_9_formula.R
\name{hr_9_formula}
\alias{hr_9_formula}
\title{Homerun / 9 formula}
\usage{
hr_9_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
HR / 9 from data
}
\description{
Calculating Homerun / 9 from data
}
\examples{
hr_9_formula(hanhwa_pitcher_2018)
}
