% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homerun_9_formula.R
\name{hr_9_formula}
\alias{hr_9_formula}
\title{Homerun / 9 calculate function}
\usage{
hr_9_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
data's HR / 9
}
\description{
Calculates Homerun / 9 in the data.
}
\examples{
hr_9_formula(find_team(hanhwa_pitcher_2018,"한화"))
}
