% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intentional_base_on_balls_formula.R
\name{ibb_formula}
\alias{ibb_formula}
\title{Intentional base on ball calculate function}
\usage{
ibb_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
The intentional base on balls number of the player
}
\description{
Calculates intentional base on balls in the data.
}
\examples{
## internal function
#ibb_formula(hanhwa_batter_2018)
}
