% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innings_pitched_formula.R
\name{ip_formula}
\alias{ip_formula}
\title{Innings pitched calculate function}
\usage{
ip_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player IP
}
\description{
Calculate IP of player
}
\examples{
## internal function
#ip_formula(hanhwa_pitcher_2018)
}
