% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2batterdata.R
\name{json2batterdf}
\alias{json2batterdf}
\title{Method of json2batterdf function}
\usage{
json2batterdf(jsondata, gameid)
}
\arguments{
\item{jsondata}{KBO game data json file you can take KBO game datas json file in github https://github.com/LOPES-HUFS/KBO_Data_Wrangling}

\item{gameid}{gameid is string type param. for example "20180324_HHWO0", 20180324 is date. HH and WO is awayteam and hometeam abbreviation. The number 0 is not a Dobleheader game. If the game is doubleheaded, the numbers are displayed as 1,2.}
}
\value{
Data with combined away team and home team data
}
\description{
The json2batterdf function will make the home team and the away team of json data a single data.
}
