% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondata2batterdf.R
\name{json2batterdf}
\alias{json2batterdf}
\title{json file to batter dataframe}
\usage{
json2batterdf(jsondata, gameid)
}
\arguments{
\item{jsondata}{KBO data json file}

\item{gameid}{KBO game date and versus infomation}
}
\value{
the batter dataframe data that is changed from json data file
}
\description{
This function changes batter data in json file to dataframe data
}
\details{
## You can read the json file in the json_sample folder and use it.
## You can use fromJSON function of jsonlite package to read the json file.
## You can also get KBO full game json file.
## You can extract it from the "https://github.com/LOPES-HUFS/KBO_Data_Wrangling" project
## And work the code in the notebook file.
## If you got a json file through one of the above methods,
## you can open it and use it as an paremeter of the function.
## The gameid paremeter can be obtained through using the default function names.
}
\examples{
## not run
# jsondata <- jsonlite::fromJSON(./json_sample/Hanhwa_normalseason_2018.json)
# json2batterdf(jsondata,names(jsondata)[1])
}
