% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsondata2pitcherdf.R
\name{json2pitcherdf}
\alias{json2pitcherdf}
\title{json file to pitcher dataframe}
\usage{
json2pitcherdf(jsondata, gameid)
}
\arguments{
\item{jsondata}{KBO data json file}

\item{gameid}{KBO game date and versus infomation}
}
\value{
A game consists of a pitcher data frame.
}
\description{
A function that creates a json file as a pitcher data frame.
}
\examples{
## You can read the json file in the json_sample folder and use it.
## You can use the fromJSON function of the jsonlite package to read the json file.
## You can also get the KBO full game json file.
## You can import it from the "https://github.com/LOPES-HUFS/KBO_Data_Wrangling" project.
## And working the code in the notebook file.
## If you get a json file through one of the above methods,
## you can open it and use it as an argument to the function.
## The gameid argument can be obtained using the default function names.
## jsondata <- jsonlite::fromJSON(./json_sample/Hanhwa_normalseason_2018.json)
# json2pitcherdf(jsondata,names(jsondata)[1])
}
