% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_9_formula.R
\name{k_9_formula}
\alias{k_9_formula}
\title{K/9 calculate function}
\usage{
k_9_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player K/9
}
\description{
Calculate K/9 of player
}
\examples{
## internal function
#k_9_formula(hanhwa_pitcher_2018)
}
