% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreboard.R
\name{make_scoreboard}
\alias{make_scoreboard}
\title{Make scoreboarddataframe}
\usage{
make_scoreboard(jsondata, gameid)
}
\arguments{
\item{jsondata}{KBO data json file}

\item{gameid}{KBO game date and versus infomation}
}
\value{
the baseball match of gameid appears in scoreboard dataframe.
}
\description{
It shows the scoreboard of the KBO game.
}
\examples{
## not run
## You can read the json file in the json_sample folder and use it.
## You can use the fromJSON function of the jsonlite package to read the json file.
## You can also get the KBO full game json file.
## You can import it from the "https://github.com/LOPES-HUFS/KBO_Data_Wrangling" project.
## jsondata <- jsonlite::fromJSON(./json_sample/Hanhwa_normalseason_2018.json)
# make_scoreboard(jsondata,names(jsondata)[1])
}
