% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_test.R
\name{name_test}
\alias{name_test}
\title{Check samename}
\usage{
name_test(name)
}
\arguments{
\item{name}{KBO player name}
}
\value{
check of samename
}
\description{
This function checks whether the name is the same or not.
}
\examples{
## internal function
name_test("이병규")
}
