% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opponent_batter_average_formula.R
\name{oba_formula}
\alias{oba_formula}
\title{Opponent batter average calculate function}
\usage{
oba_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
data's OBA
}
\description{
Calculate OBA in the data
}
\examples{
oba_formula(hanhwa_pitcher_2018[hanhwa_pitcher_2018$team=="한화",])
}
