% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_base_percentage_formula.R
\name{obp_formula}
\alias{obp_formula}
\title{On-base Percentage formula}
\usage{
obp_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
OBP from data
}
\description{
Calculating OBP from data
}
\examples{
obp_formula(hanhwa_batter_2018)
}
