% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onebase_formula.R
\name{onebase_formula}
\alias{onebase_formula}
\title{1B calculate function}
\usage{
onebase_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
the number of 1B
}
\description{
count single hit of player
}
\examples{
## internal function
#onebase_formula(hanhwa_batter_2018)
}
