% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onebase_formula.R
\name{onebase_formula}
\alias{onebase_formula}
\title{1B count}
\usage{
onebase_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
1B from data
}
\description{
Counting single hit from data.
}
\examples{
onebase_formula(hanhwa_batter_2018)
}
