% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitches_per_inning_pitched_formula.R
\name{p_ip_formula}
\alias{p_ip_formula}
\title{Pitches per inning pitched calculate function}
\usage{
p_ip_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
data's P/IP
}
\description{
Calculate P/IP in the data
}
\examples{
p_ip_formula(hanhwa_pitcher_2018)
}
