% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitches_per_inning_pitched_formula.R
\name{p_ip_formula}
\alias{p_ip_formula}
\title{Pitches Per Inning Pitched formula}
\usage{
p_ip_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
P/IP from data
}
\description{
Calculating P/IP from data
}
\examples{
p_ip_formula(hanhwa_pitcher_2018)
}
