% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_appearance_formula.R
\name{pa_formula}
\alias{pa_formula}
\title{The number of Plate Appearance formula}
\usage{
pa_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
PA from data
}
\description{
Calculating PA from data
}
\examples{
pa_formula(hanhwa_batter_2018)
}
