% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_SO.R
\name{player_SO}
\alias{player_SO}
\title{Calculating batter's strikeout}
\usage{
player_SO(data = kbodatamining::hanhwa_batter_2018, playername,
  yearly = NULL, monthly = NULL)
}
\arguments{
\item{data}{Default data is hanhwa_batter_2018 You can put other kbo batter data if you want.}

\item{playername}{The name of the batter you want}

\item{yearly}{The default value is NULL Possible values are a specific year.}

\item{monthly}{The default value is NULL Possible values are a specific month.}
}
\value{
The strikeout for the year, month, or period put in is calculated.
}
\description{
This function calculates the strikeout for the specified player.
This function has an option for yearly and monthly.
default: the annual strikeout is calculated
If you put a year in the yearly parameter, it calculates the corresponding year strikeout.
If you put a month in the monthly parameter, it calculates the corresponding yearly-month strikeout.
If you put a year and a month in two parameters, the strikeout for the specific time you put in is calculated.
}
\examples{
## default
player_SO(hanhwa_batter_2018,"이용규")
## yearly=2018, monthly=NULL
player_SO(hanhwa_batter_2018,"이용규",yearly=2018)
## yearly=NULL, monthly="05"
player_SO(hanhwa_batter_2018,"이용규",monthly="05")
## yearly=2018, monthly="05"
player_SO(hanhwa_batter_2018,"이용규",yearly=2018,monthly="05")
}
