% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreboard.R
\name{scoreboard}
\alias{scoreboard}
\title{Make scoreboard dataframe}
\usage{
scoreboard(jsondata, gameid)
}
\arguments{
\item{jsondata}{You can try the function by importing the file from the json_sample folder. Alternatively, you can create a json-formatted KBO data file by referring to the Readme file.}

\item{gameid}{KBO game date and versus infomation, You can get this information by importing the name of the json file.}
}
\value{
gameid's baseball game appears as a scoreboard data frame.
}
\description{
This function creates a scoreboard for that information based on json data and gameid.
}
\examples{
## not run
# jsondata <- jsonlite::fromJSON("./json_sample/Hanhwa_normalseason_2018.json")
# scoreboard(jsondata,names(jsondata)[1])
}
