% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacrifice_fly_formula.R
\name{sf_formula}
\alias{sf_formula}
\title{Sacrifice fly calculate function}
\usage{
sf_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's Sacrifice fly
}
\description{
Calculate Sacrifice fly in the data
}
\examples{
sf_formula(hanhwa_batter_2018)
}
