% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacrifice_fly_formula.R
\name{sf_formula}
\alias{sf_formula}
\title{Sacrifice Fly formula}
\usage{
sf_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
Sacrifice Fly from data
}
\description{
Calculating Sacrifice Fly from data
}
\examples{
sf_formula(hanhwa_batter_2018)
}
