% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacrifice_hit_formula.R
\name{sh_formula}
\alias{sh_formula}
\title{Sacrifice hit calculate function}
\usage{
sh_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's Sacrifice hit
}
\description{
Calculate Sacrifice hit in the data
}
\examples{
sh_formula(hanhwa_batter_2018)
}
