% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugging_average.R
\name{slg_formula}
\alias{slg_formula}
\title{Slugging average calculate function}
\usage{
slg_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
player Slugging average
}
\description{
Calculate Slugging average of player
}
\examples{
## internal function
#obp_formula(hanhwa_batter_2018)
}
