% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugging_average_formula.R
\name{slg_formula}
\alias{slg_formula}
\title{Slugging Average formula}
\usage{
slg_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
Slugging Average from data
}
\description{
Calculating Slugging Average from data
}
\examples{
slg_formula(hanhwa_batter_2018)
}
