% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/so_formula.R
\name{so_formula}
\alias{so_formula}
\title{SO calculate formula}
\usage{
so_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
count SO
}
\description{
SO calculate function
}
\details{
count SO of player
}
\examples{
## internal function
#so_formula(hanhwa_batter_2018)
}
