% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_batter_compare_plot.R
\name{team_batter_compare_plot}
\alias{team_batter_compare_plot}
\title{Comparing the hits of two teams}
\usage{
team_batter_compare_plot(data)
}
\arguments{
\item{data}{two teams of batter's merged data}
}
\value{
A plot comparing records(AVG,OBP,SLG,BABIP)
}
\description{
A plotting function that compares the hits of two teams
We compare batters who have played in more than ten games on each team.
}
\examples{
team_batter_compare_plot(make_all_batter_data(hanhwa_batter_2018,"한화","LG",2018,NULL))
}
