% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_disposal.R
\name{team_disposal}
\alias{team_disposal}
\title{Process the team's complete game record}
\usage{
team_disposal(data, boxscore, teamname)
}
\arguments{
\item{data}{cg_list or cgs_list (The result value of bind_cg_list or bind_cgs_list)}

\item{boxscore}{The result value of the pitcher_statistics function}

\item{teamname}{teamname}
}
\value{
CG or CGS
}
\description{
Process for creating team box scores
}
\examples{
## internal function
## not rn
#team_dispose(cg_list,boxscore)
}
