% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_pitcher_compare_plot.R
\name{team_pitcher_compare_plot}
\alias{team_pitcher_compare_plot}
\title{Comparing the pitching of two teams}
\usage{
team_pitcher_compare_plot(data)
}
\arguments{
\item{data}{data of all pitchers from two teams}
}
\value{
A plot comparing records(ERA,K(SO),OBA,P/IP)
}
\description{
A plotting function that compares the pitching of two teams.
}
\examples{
team_pitcher_compare_plot(make_all_pitcher_data(hanhwa_pitcher_2018,"한화","LG",2018,NULL))
}
