% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threebase_formula.R
\name{threebase_formula}
\alias{threebase_formula}
\title{3B count}
\usage{
threebase_formula(data)
}
\arguments{
\item{data}{KBO batter data}
}
\value{
3B from data
}
\description{
Counting Triples from data
}
\examples{
threebase_formula(hanhwa_batter_2018)
}
