% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twobase_formula.R
\name{twobase_formula}
\alias{twobase_formula}
\title{2B calculate function}
\usage{
twobase_formula(data)
}
\arguments{
\item{data}{batter inning data}
}
\value{
data's 2B
}
\description{
Calculates the data's Double
}
\examples{
twobase_formula(hanhwa_batter_2018)
}
