% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winning_percentage_formula.R
\name{wpct_formula}
\alias{wpct_formula}
\title{A pitcher's winning percentage calculate function}
\usage{
wpct_formula(data)
}
\arguments{
\item{data}{KBO pitcher data}
}
\value{
player WPCT
}
\description{
Calculate WPCT of player
}
\examples{
## internal function
#wpct_formula(hanhwa_pitcher_2018)
}
