% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_batter_plot.R
\name{yearly_batter_plot}
\alias{yearly_batter_plot}
\title{Annual hit record plot}
\usage{
yearly_batter_plot(data)
}
\arguments{
\item{data}{You can use the `make_batter_data_4plot()` function to create merged data of two batters'.}
}
\value{
Four plot comparing four records(G,AVG,OPS,BABIP) of each player
}
\description{
This function compares the four records of each two players from the data.
This function requires data that have a duration of at least two seasons.
}
\examples{
## internal function
#yearly_batter_plot(make_batter_data_4plot(Yearly_KBO_DATA,"이용규","정근우"))
}
