% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_month_record.R
\name{yearly_month_record}
\alias{yearly_month_record}
\title{yearly month record show}
\usage{
yearly_month_record(data, playername, recordname, month)
}
\arguments{
\item{data}{default is hanhwa's batter data, You can include the entire KBO data if you wish.}

\item{playername}{The name of the batter you want}

\item{recordname}{The name of the record you want}

\item{month}{The month you want to find}
}
\value{
show yearly-month batter record
}
\description{
It shows the yearly-month batter record.
}
\examples{
## internal function
#yearly_month_record(hanhwa_batter_2018,"이용규",avg_formula)
}
