% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_month_record.R
\name{yearly_month_record}
\alias{yearly_month_record}
\title{Monthly record of every year}
\usage{
yearly_month_record(data, name, id, recordname, month)
}
\arguments{
\item{data}{KBO datas with at least one year of data}

\item{name}{The name of the player or the team you want}

\item{id}{The id of the player you want}

\item{recordname}{The name of the record you want}

\item{month}{Possible value is a specific month}
}
\value{
the Monthly record of every year of the player or team
}
\description{
This function shows the Monthly record of every year of the player or team
}
\examples{
## internal function - not run
#yearly_month_record(hanhwa_batter_2018,"이용규",avg_formula,"05")
#yearly_month_record(hanhwa_batter_2018,"한화",avg_formula,"05")
}
