% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_pitcher_plot.R
\name{yearly_pitcher_plot}
\alias{yearly_pitcher_plot}
\title{Yearly pitcher data plotting}
\usage{
yearly_pitcher_plot(data)
}
\arguments{
\item{data}{two pitcher player's merged data}
}
\value{
A plot comparing four records(G,ERA,K,IP)
}
\description{
plotting yearly pitcher data
}
\examples{
## internal function
#yearly_pitcher_plot(make_pitcher_data_4plot(Yearly_KBO_DATA,"정우람","이태양"))
}
