% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_record.R
\name{yearly_record}
\alias{yearly_record}
\title{Yearly record show}
\usage{
yearly_record(data, playername, recordname)
}
\arguments{
\item{data}{default is hanhwa's batter data, You can include the entire KBO data if you wish.}

\item{playername}{The name of the batter you want}

\item{recordname}{The name of the record you want}
}
\value{
calculate yearly batter record
}
\description{
It shows the yearly batter record
}
\examples{
## internal function
#yearly_record(hanhwa_batter_2018,"이용규",avg_formula)
}
